rename files sensibly, fix styling of new record dialog
