+++ appv2/src/UI/Map2/Controls/CenterCurrentRecord.tsx
